fixed properties bug
