+++ src/test/java/bio/terra/cda/app/service/QueryServiceTest.java
@@ -16,7 +16,9 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.springframework.stereotype.Component;
 
+@Component
 class QueryServiceTest {
 
   private final QueryService queryService = new QueryService(new ObjectMapper());
